Fix several consistent update issues
ConsistentUpdate was inefficiently retrying by loading the
data at the end of a failed attempt, rather than at the beggining
of the next attempt, causing many more dirty reads
in highly concurrent scenarious and exhausting retries that
would otherwise succeed.

Also the in memory svc implementation was not serializing access
to data correctly, which was causing silent ConsistentUpdate
failures due to stale reads.