Add license CLI
There's nothing out there that really works well. The only
that gets close is https://github.com/lluissm/license-header-checker
but it's trying to do too many things (takes a directory, then you need
to pass file extensions and or sub directories that you want to filter out),
and a simple "Check all .go files but exclude .pb.go" is not possible to do.

This CLI takes a list of files, which means that the aforementioned operation is
as easy as `bluectl license LICENSE $(find . -name \*.go | grep -v .pb.go | xargs)`.