Update firstmover.Service to not remove lockFile
When grpc server is done, there's a race between leader to remove lock
and a follower to take over leader, creating a lock file. If the follower
goes first, the socket is removed from under it, causing a potential split brain,
as a new follower is incapable of connecting to the newly upgraded leader.