Update iterator.IsEmpty to Close when empty
The way this function is tipically used is by overriding the variable
holding the passed iterator. Before this change it was not apparent
what happened to the consumed iterator in the case it's empty.

Now the passed iterator is consumed either way, so the caller doesn't
need to worry about Close not being called.